// types/zod/outputTypeSchemas/usersFindManyArgsSchema.ts
export const usersFindManyArgsSchema: z.ZodType<Prisma.usersFindManyArgs> = z
	.object({
		select: usersSelectSchema.optional(),
		include: usersIncludeSchema.optional(),
		where: usersWhereInputSchema.optional(),
		orderBy: z.union([usersOrderByWithRelationInputSchema.array(), usersOrderByWithRelationInputSchema]).optional(),
		cursor: usersWhereUniqueInputSchema.optional(),
		take: z.number().optional(),
		skip: z.number().optional(),
		distinct: z.union([UsersScalarFieldEnumSchema, UsersScalarFieldEnumSchema.array()]).optional(),
	})
	.strict();